requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
grpcio==1.74.0
flask==3.0.0
flask-cors==4.0.0 